requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
filelock==3.11.0
Flask==2.2.3
idna==3.4
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
Pillow==9.5.0
requests==2.28.2
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
